+++ test/tests/warnings.test.ts
@@ -1,34 +1,12 @@
 import { logger } from '../../src/logSettings';
-import { arrayProp, buildSchema, mapProp, prop } from '../../src/typegoose';
+import { arrayProp, buildSchema, mapProp } from '../../src/typegoose';
 
 beforeEach(() => {
   logger.warn = jest.fn();
   expect.assertions(1);
 });
 
-// expect((logger.warn as any).mock.calls.length).toEqual(1);
-
 describe('prop.ts', () => {
-  describe('@prop', () => {
-    it('should warn if option "items" is used in an @prop', () => {
-      class TestPropOptionItems {
-        @prop({ items: String })
-        public test: any[];
-      }
-      buildSchema(TestPropOptionItems);
-      expect((logger.warn as any).mock.calls.length).toEqual(1);
-    });
-
-    it('should warn if option "of" is used in an @prop', () => {
-      class TestPropOptionOf {
-        @prop({ of: String })
-        public test: any[];
-      }
-      buildSchema(TestPropOptionOf);
-      expect((logger.warn as any).mock.calls.length).toEqual(1);
-    });
-  });
-
   describe('@mapProp', () => {
     it('should warn if option "items" is used in an @mapProp', () => {
       class TestMapPropOptionItems {